backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy>=1.21.0
python-multipart==0.0.6
typing-extensions==4.8.0
python-dateutil==2.8.2
requests==2.32.3